requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
# LangChain ecosystem (compatible versions)
langchain-core==0.2.38
langchain==0.2.16
groq==0.29.0
langgraph==0.2.16
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.2
google-auth==2.40.3
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.32.4
python-dateutil==2.8.2
pytz==2023.3
pytest==7.4.3
cachetools==5.5.2
rsa==4.9.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
requests-oauthlib==2.0.0
oauthlib==3.3.1
charset-normalizer==3.4.2
urllib3==2.5.0